moving cli/contexts/project/diff.py to cli/util/compare.py
