+++ src/datumaro/cli/commands/diff.py
@@ -8,13 +8,13 @@ import os
 import os.path as osp
 from enum import Enum, auto
 
+from datumaro.cli.util.compare import DiffVisualizer
 from datumaro.components.comparator import DistanceComparator, EqualityComparator
 from datumaro.components.errors import ProjectNotFoundError
 from datumaro.util import dump_json_file
 from datumaro.util.os_util import rmtree
 from datumaro.util.scope import on_error_do, scope_add, scoped
 
-from ..contexts.project.diff import DiffVisualizer
 from ..util import MultilineFormatter
 from ..util.errors import CliException
 from ..util.project import generate_next_file_name, load_project, parse_full_revpath
