+++ src/datumaro/cli/util/compare.py
